requirements.txt
Django==1.5.1
django-annoying==0.7.7
